+++ app_update_checker/lib/src/controller/update_controller.dart
@@ -5,16 +5,16 @@ import 'dart:async';
 import 'package:flutter/foundation.dart';
 import 'package:package_info_plus/package_info_plus.dart';
 
-import '../builder/models/app_update.dart';
-import '../builder/models/release.dart';
-import '../builder/models/update_config.dart';
-import '../config/models/release_settings_config.dart';
-import '../config/models/store_config.dart';
-import '../config/update_config_parser.dart';
+import '../localizer/models/app_update.dart';
+import '../localizer/models/release.dart';
+import '../localizer/models/update_config.dart';
+import '../parser/models/release_settings_config.dart';
+import '../parser/models/store_config.dart';
+import '../parser/update_config_parser.dart';
 import '../linker/models/release_data.dart';
 import '../linker/models/update_config_data.dart';
 import '../linker/update_config_linker.dart';
-import '../models/update_platform.dart';
+import '../shared/update_platform.dart';
 import '../stores/fetchers/store_fetcher.dart';
 import 'update_config_provider.dart';
 import 'update_contoller_base.dart';
