+++ app_update_checker/lib/src/finder/update_finder.dart
@@ -1,10 +1,10 @@
 // ignore_for_file: avoid-unnecessary-reassignment, avoid-nested-switches, prefer-correct-identifier-length
 
-import '../builder/models/release.dart';
-import '../builder/models/update_config.dart';
-import '../models/release_status.dart';
-import '../models/update_platform.dart';
-import '../models/version.dart';
+import '../localizer/models/release.dart';
+import '../localizer/models/update_config.dart';
+import '../shared/release_status.dart';
+import '../shared/update_platform.dart';
+import '../shared/version.dart';
 
 class UpdateFinder {
   final Version appVersion;
