+++ app_update_checker/lib/src/linker/models/release_settings.dart
@@ -1,8 +1,8 @@
 import 'dart:ui';
 
-import '../../config/models/release_settings_config.dart';
-import '../../models/text_translations.dart';
-import '../../models/version.dart';
+import '../../parser/models/release_settings_config.dart';
+import '../../shared/text_translations.dart';
+import '../../shared/version.dart';
 
 class ReleaseSettings {
   final TextTranslations titleTranslations;
