+++ app_update_checker/lib/src/parser/models/release_settings_config.dart
@@ -1,5 +1,5 @@
-import '../../models/text_translations.dart';
-import '../../models/version.dart';
+import '../../shared/text_translations.dart';
+import '../../shared/version.dart';
 
 class ReleaseSettingsConfig {
   final TextTranslations? titleTranslations;
