+++ app_update_checker/lib/src/parser/models/update_config_exception.dart
