+++ app_update_checker/lib/src/parser/sub_parsers/duration_parser.dart
