+++ app_update_checker/lib/src/parser/sub_parsers/release_parser.dart
