+++ app_update_checker/lib/src/parser/sub_parsers/store_parser.dart
