+++ app_update_checker/lib/src/parser/sub_parsers/text_parser.dart
