+++ app_update_checker/lib/src/parser/sub_parsers/version_parser.dart
