+++ app_update_checker/lib/src/shared/release_status.dart
