+++ app_update_checker/lib/src/shared/text_translations.dart
