+++ app_update_checker/lib/src/shared/update_platform.dart
