+++ app_update_checker/lib/src/shared/version.dart
