+++ app_update_checker/lib/src/stores/fetchers/store_fetcher.dart
@@ -3,7 +3,7 @@ import 'dart:ui';
 
 import 'package:package_info_plus/package_info_plus.dart';
 
-import '../../config/models/release_config.dart';
+import '../../parser/models/release_config.dart';
 import '../store.dart';
 import 'app_store_fetcher.dart';
 import 'google_play_fetcher.dart';
