+++ app_update_checker/lib/src/stores/store.dart
@@ -1,4 +1,4 @@
-import '../models/update_platform.dart';
+import '../shared/update_platform.dart';
 
 enum Stores {
   googlePlay,
